backend update

+++ backend/api/views.py
@@ -3,9 +3,13 @@ from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from django.contrib.auth import authenticate
 from rest_framework.authtoken.models import Token
+from django.contrib.auth.models import User
+from rest_framework import status
+from django.shortcuts import get_object_or_404
+from rest_framework.pagination import PageNumberPagination
 
 from django.db.models import Avg
-from .models import Movie, Review, Rating,  Actor, CastMovie
+from .models import Movie, Review, Rating,  Actor, CastMovie, Recommendation, Watchlist
 from .serializers import (
     MovieSerializer,
     MovieCreateSerializer,
@@ -13,7 +17,11 @@ from .serializers import (
     LoginSerializer,
     RegisterSerializer,
     ActorSerializer, 
-    CastMovieSerializer
+    CastMovieSerializer,
+    WatchlistSerializer,
+    RecommendationSerializer,
+    CastMovieCreateSerializer,
+    RecommendationCreateSerializer
 )
 
 
@@ -29,7 +37,7 @@ class LoginAPIView(APIView):
             token, _ = Token.objects.get_or_create(user=user)
             return Response({'token': token.key})
 
-        return Response({'error': 'Invalid credentials'})
+        return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
     
 
 # REGISTER
@@ -47,13 +55,27 @@ class RegisterAPIView(APIView):
 class MovieListAPIView(APIView):
     def get(self, request):
         movies = Movie.objects.all()
-        serializer = MovieSerializer(movies, many=True)
-        return Response(serializer.data)
+        
+        search = request.GET.get('search')
+        genre = request.GET.get('genre')
+
+        if search:
+            movies = movies.filter(title__icontains=search)
+
+        if genre:
+            movies = movies.filter(moviegenre__genre__name__icontains=genre)
+            
+        paginator = PageNumberPagination()
+        paginator.page_size = 5 
+        
+        result_page = paginator.paginate_queryset(movies, request)   
+        serializer = MovieSerializer(result_page, many=True)
+        return paginator.get_paginated_response(serializer.data)
     
 # GET ONE MOVIE
 class MovieDetailAPIView(APIView):
     def get(self, request, pk):
-        movie = Movie.objects.get(id=pk)
+        movie = get_object_or_404(Movie, id=pk)
         serializer = MovieSerializer(movie)
         return Response(serializer.data)
 
@@ -62,6 +84,10 @@ class MovieCreateAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request):
+        
+        if not request.user.is_staff:
+            return Response({"error": "Only admin can add movies"}, status=403)
+        
         serializer = MovieCreateSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -73,7 +99,11 @@ class MovieUpdateAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
     def put(self, request, pk):
-        movie = Movie.objects.get(id=pk)
+        
+        if not request.user.is_staff:
+            return Response({"error": "Only admin can update movies"}, status=403)
+        
+        movie = get_object_or_404(Movie, id=pk)
         serializer = MovieCreateSerializer(movie, data=request.data)
 
         if serializer.is_valid():
@@ -86,7 +116,11 @@ class MovieDeleteAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
     def delete(self, request, pk):
-        movie = Movie.objects.get(id=pk)
+        
+        if not request.user.is_staff:
+            return Response({"error": "Only admin can delete movies"}, status=403)
+        
+        movie = get_object_or_404(Movie, id=pk)
         movie.delete()
         return Response({'message': 'Deleted'})
 
@@ -147,6 +181,8 @@ class ActorListAPIView(APIView):
         return Response(serializer.data)
     
 class CastMovieAPIView(APIView):
+    
+    permission_classes = [IsAuthenticated]
 
     def get(self, request, movie_id):
         cast = CastMovie.objects.filter(movie_id=movie_id)
@@ -154,17 +190,83 @@ class CastMovieAPIView(APIView):
         return Response(serializer.data)
 
     def post(self, request, movie_id):
+        if not request.user.is_staff:
+           return Response({"error": "Admin Only"}, status=status.HTTP_403_FORBIDDEN)
+        serializer = CastMovieCreateSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
         actor_id = request.data.get('actor_id')
         role_name = request.data.get('role_name', '')
 
-        if not actor_id:
-            return Response({'error': 'actor_id is required'})
+        actor = get_object_or_404(Actor, id=actor_id)
+        movie = get_object_or_404(Movie, id=movie_id)
 
         cast, created = CastMovie.objects.get_or_create(
             movie_id=movie_id,
             actor_id=actor_id,
             defaults={'role_name': role_name}
         )
-
-        serializer = CastMovieSerializer(cast)
+        return Response(CastMovieSerializer(cast).data)
+    
+class WatchlistAPIView(APIView):
+    permission_classes=[IsAuthenticated]
+    
+    def get(self,request):
+        watchlist = Watchlist.objects.filter(user=request.user).select_related('movie')
+        serializer = MovieSerializer(
+            [item.movie for item in watchlist], many=True
+        )
         return Response(serializer.data)
+    def post(self,request):
+        movie_id=request.data.get('movie')
+        
+        obj, created=Watchlist.objects.get_or_create(
+            user=request.user,
+            movie_id=movie_id
+        )   
+        
+        if not created:
+            return Response({'message':'Already in watchlist'},status=status.HTTP_200_OK)    
+        
+        return Response({'message':'Added to watchlist'},status=status.HTTP_201_CREATED)
+    def delete(self,request):
+        movie_id = request.data.get('movie')
+        Watchlist.objects.filter(user=request.user, movie_id=movie_id).delete()
+        return Response({'message': 'Removed'})
+        
+class RecommendationAPIView(APIView):
+    permission_classes=[IsAuthenticated]
+    
+    def post(self,request):
+        serializer = RecommendationCreateSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        
+        data = serializer.validated_data
+        to_username=data['to_username']
+        movie_title=data['movie_title']
+        message=data.get('message', '') 
+        
+        to_user = get_object_or_404(User, username=to_username)
+        if to_user==request.user:
+            return Response({'error': 'You can not recommend yourself'}, status=status.HTTP_400_BAD_REQUEST)
+            
+        
+        movie = get_object_or_404(Movie, title=movie_title)
+          
+        Recommendation.objects.create(
+            from_user=request.user,
+            to_user=to_user,
+            movie=movie,
+            message=message
+        )
+
+        return Response({'message': 'Recommendation sent'},status=status.HTTP_201_CREATED)     
+      
+class RecommendationListAPIView(APIView): 
+    permission_classes = [IsAuthenticated]
+    
+    def get(self, request):
+        recs = Recommendation.objects.filter(to_user=request.user)
+        serializer = RecommendationSerializer(recs, many=True)
+        return Response(serializer.data)
+      
+
